requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29
uvicorn
fastapi
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.4.1
aiosignal==1.2.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
dataclasses==0.6
elasticsearch==7.15.2
et-xmlfile==1.1.0
fake-useragent==0.1.11
fastapi==0.70.0
frozenlist==1.2.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
h11==0.12.0
idna==3.3
multidict==5.2.0
oauthlib==3.1.1
openai==0.11.0
openpyxl==3.0.9
pandas-stubs==1.2.0.38
pycares==4.1.2
pycparser==2.21
pydantic==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-socks==1.2.4
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.16.0
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
yarl==1.7.2

# data science
numpy
pandas

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
